requirements.txt
# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Security testing
bandit>=1.7.0
safety>=2.0.0

# Code quality
flake8>=5.0.0
black>=22.0.0

# Additional DevSecOps tools
pyyaml>=6.0
requests>=2.28.0

# Include demo-sbom-lab requirements
-r solutions/demo-sbom-lab/requirements.txt
